+++ packages/implementations/react/src/atomic/Badge/src/index.ts
@@ -1,5 +1,5 @@
 import { Badge } from "./Badge";
 
 export { Badge } from "./Badge";
-export type { BadgeProps } from "./badge.types";
+export type { BadgeProps, BadgeProperties } from "./badge.types";
 export default Badge;