+++ packages/implementations/react/src/composite/Stepper/src/components/Step/index.ts
@@ -1,0 +1,6 @@
+import { Step } from "./Step";
+
+export { Step } from "./Step";
+export type { StepProps } from "./step.types";
+
+export default Step;