+++ packages/implementations/react/src/composite/Stepper/src/context/Manager/index.ts
@@ -1,0 +1,5 @@
+import { Context } from "./Manager";
+
+export { Context } from "./Manager";
+export type { ContextProps } from "./manager.types";
+export default Context;