+++ packages/implementations/react/src/composite/Stepper/src/index.ts
@@ -1,0 +1,5 @@
+import { Stepper } from "./Stepper";
+
+export { Stepper } from "./Stepper";
+export type { StepperProps } from "./stepper.types";
+export default Stepper;