+++ packages/styles/src/packages/atomic/stepper/index.ts
@@ -1,0 +1,8 @@
+import { stepperStyles } from "./rarui-stepper.css";
+
+export const stepper = { classNames: stepperStyles };
+
+export type {
+  StepperVariants,
+  StepperStepVariants,
+} from "./rarui-stepper.types";