+++ packages/styles/src/packages/atomic/stepper/rarui-stepper.types.ts
@@ -1,0 +1,7 @@
+import { RecipeVariants } from "@vanilla-extract/recipes";
+import { stepperStyles } from "./rarui-stepper.css";
+
+export type StepperVariants = RecipeVariants<typeof stepperStyles.stepper>;
+export type StepperStepVariants = RecipeVariants<
+  typeof stepperStyles.stepCircle
+>;